Ability to round up/down in multiply and divide operations (#13)
The default behaviour for the multiply and divide operations was to ROUND_HALF.
Added the possibility to specify whether to ROUND_UP or ROUND_DOWN. Rounding
down is especially important as that is the typical behaviour in smart contract
opcodes (EVM and non-EVM alike).

Also fixed a bug in the divideAndRound function (which is called by many
operations) when the divisor is negative. E.g. divideAndRound(-15n, 2n)) and
divideAndRound(15n, -2n) yielded different results.

Updated docs with latest changes.